将商业版本ServiceStack.Redis 替换成 StackExchange 重写 redisCache类

+++ Learun.Framework.Module/Learun.Cache/Learun.Cache.Redis/CacheByRedis.cs
@@ -19,7 +19,7 @@ namespace Learun.Cache.Redis
         /// </summary>
         /// <param name="cacheKey">键</param>
         /// <returns></returns>
-        public T Read<T>(string cacheKey, long dbId = 0) where T : class
+        public T Read<T>(string cacheKey, int dbId=0) where T : class
         {
             return RedisCache.Get<T>(cacheKey, dbId);
         }
@@ -28,7 +28,7 @@ namespace Learun.Cache.Redis
         /// </summary>
         /// <param name="value">对象数据</param>
         /// <param name="cacheKey">键</param>
-        public void Write<T>(string cacheKey, T value, long dbId = 0) where T : class
+        public void Write<T>(string cacheKey, T value, int dbId=0) where T : class
         {
             RedisCache.Set(cacheKey, value, dbId);
         }
@@ -38,7 +38,7 @@ namespace Learun.Cache.Redis
         /// <param name="value">对象数据</param>
         /// <param name="cacheKey">键</param>
         /// <param name="expireTime">到期时间</param>
-        public void Write<T>(string cacheKey, T value, DateTime expireTime, long dbId = 0) where T : class
+        public void Write<T>(string cacheKey, T value, DateTime expireTime, int dbId=0) where T : class
         {
             RedisCache.Set(cacheKey, value, expireTime, dbId);
         }
@@ -48,7 +48,7 @@ namespace Learun.Cache.Redis
         /// <param name="value">对象数据</param>
         /// <param name="cacheKey">键</param>
         /// <param name="TimeSpan">缓存时间</param>
-        public void Write<T>(string cacheKey, T value, TimeSpan timeSpan, long dbId = 0) where T : class
+        public void Write<T>(string cacheKey, T value, TimeSpan timeSpan, int dbId=0) where T : class
         {
             RedisCache.Set(cacheKey, value, timeSpan, dbId);
         }
@@ -56,14 +56,14 @@ namespace Learun.Cache.Redis
         /// 移除指定数据缓存
         /// </summary>
         /// <param name="cacheKey">键</param>
-        public void Remove(string cacheKey, long dbId = 0)
+        public void Remove(string cacheKey, int dbId=0)
         {
             RedisCache.Remove(cacheKey, dbId);
         }
         /// <summary>
         /// 移除全部缓存
         /// </summary>
-        public void RemoveAll(long dbId = 0)
+        public void RemoveAll(int dbId=0)
         {
             RedisCache.RemoveAll(dbId);
         }
